my_tgbot/requirements.txt
requests
psycopg2-binary
telebot